Section 33. Putting Everything Together finish

+++ 28.todolist-v2/app.js
@@ -1,6 +1,8 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
+
 const app = express();
 
 app.set("view engine", "ejs");
@@ -40,7 +42,7 @@ const listSchema = {
 
 const List = mongoose.model("List", listSchema);
 
-Item.insertMany(defaultItems);
+//Item.insertMany(defaultItems);
 
 app.get("/", function (req, res) {
   awaitFind();
@@ -61,7 +63,7 @@ app.get("/", function (req, res) {
 });
 
 app.get("/:customListName", function (req, res) {
-  const customListName = req.params.customListName;
+  const customListName = _.capitalize(req.params.customListName);
 
   List.findOne({ name: customListName })
     .then(function (foundList) {
@@ -110,8 +112,20 @@ app.post("/", function (req, res) {
 
 app.post("/delete", function (req, res) {
   const checkedItemId = req.body.checkbox;
-  Item.findByIdAndRemove(checkedItemId).then(function (err) {});
-  res.redirect("/");
+  const listName = req.body.listName;
+
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId).then(function (err) {});
+    res.redirect("/");
+  } else {
+    List.findOneAndUpdate(
+      { name: listName },
+      { $pull: { items: { _id: checkedItemId } } }
+    ).then(function (err) {
+      res.redirect("/" + listName);
+      console.log(err);
+    });
+  }
 });
 
 app.get("/work", function (req, res) {
